moving benchmark tests into their own file to prevent timeout failures of other tests potentially stopping the benchmarks
